+++ src/libintervalxt/label.cc
@@ -36,6 +36,6 @@ Label<Tlen, Tmat>::Label() {
 // Explicit instantiations of templates so that code is generated for the linker.
 #include <gmpxx.h>
 
-template class intervalxt::Label<unsigned long, unsigned long>;
+template class intervalxt::Label<int, int>;
 template class intervalxt::Label<mpz_class, mpz_class>;
-template class intervalxt::Label<mpz_class, unsigned long>;
+template class intervalxt::Label<mpz_class, int>;
